+++ client/src/components/basic/select/select.module.scss
@@ -61,6 +61,8 @@
 	box-shadow: 0px 4px 6px rgba(189, 189, 189, 0.25);
 
 	list-style-type: none;
+	white-space: nowrap;
+	overflow: hidden;
 }
 .optionsActive {
 	display: block;
